klient: Moved klient and the list type to their own package
For repair to use
