+++ go/src/koding/klientctl/mount.go
@@ -13,7 +13,9 @@ import (
 	"time"
 
 	"koding/klient/remote/req"
+	"koding/klientctl/klient"
 	"koding/klientctl/klientctlerrors"
+	"koding/klientctl/list"
 	"koding/klientctl/util"
 
 	"github.com/cheggaaa/pb"
@@ -46,7 +48,7 @@ type MountCommand struct {
 
 	// The klient instance this struct will use.
 	Klient interface {
-		RemoteList() (KiteInfos, error)
+		RemoteList() (list.KiteInfos, error)
 		RemoteCache(req.Cache, func(par *dnode.Partial)) error
 		RemoteMountFolder(req.MountFolder) (string, error)
 
@@ -62,7 +64,7 @@ type MountCommand struct {
 	//
 	// Note! These will be ignored if c.Klient is already defined before Run() is
 	// called.
-	KlientOptions KlientOptions
+	KlientOptions klient.KlientOptions
 
 	// the following vars exist primarily for mocking ability, and ensuring
 	// an enclosed environment within the struct.
@@ -249,7 +251,7 @@ func (c *MountCommand) setupKlient() (int, error) {
 		return 0, nil
 	}
 
-	k, err := NewDialedKlient(c.KlientOptions)
+	k, err := klient.NewDialedKlient(c.KlientOptions)
 	if err != nil {
 		return 1, fmt.Errorf("Failed to get working Klient instance")
 	}
